+++ src/lib/core/elements/AbstractBaseEngine.server.ts
@@ -1,6 +1,6 @@
 import type { JsonMap, imagePath } from "$exports";
 import { resolve, extname } from "node:path";
-import { promises as fs } from "fs";
+import { promises as fs } from "node:fs";
 
 export abstract class AbstractBaseEngine {
